Implement placing ships through form
Refactored checkWinner test.

Removed some duplications in Gameboard.js
and added getNumberOfShips function to only allow
the game to be started if 5 ships are placed on
the board.

Added a helper function in DOMController.js to
remove duplications.

Added a function in GameController.js that
randomly places ships for the bot